Refactor: factorize stack operation functions

+++ libft/ft_lib/ft_lstnew.c
@@ -6,13 +6,13 @@
 /*   By: abendrih <abendrih@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/13 23:09:24 by abendrih          #+#    #+#             */
-/*   Updated: 2025/08/03 23:41:47 by abendrih         ###   ########.fr       */
+/*   Updated: 2025/08/05 03:05:55 by abendrih         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-t_stack_node	*ft_lstnew(long value)
+t_stack_node	*ft_lstnew(int value)
 {
 	t_stack_node	*key;
